passwordmanager/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
Django==5.2.4
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-formtools==2.5.1
django-otp==1.6.1
django-phonenumber-field==8.1.0
django-vite==3.1.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
frozenlist==1.7.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
multidict==6.6.3
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.2
PyJWT==2.8.0
pyotp==2.9.0
pypng==0.20220715.0
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
sqlparse==0.5.3
twilio==9.7.0
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.20.1
